Reintroduce PlayerDataStore as PDC wrapper
User prompt: Oh um hmm. maybe a datastore is the right call, the main class is already massive and messy enough. unless there's a less complex way. I assume it's messier/duplicated code if each getter had to parse out the data from the PDC itself I assume. So can probably put back the datastore, or make it more like pdc helper methods or something idk.

Caveman summary: Agree. Main class huge. Keep PDC backend. Move name color and exp storage logic into small PlayerDataStore wrapper. Main class owns instance only. Callers use datastore.
