+++ src/main/java/me/robomwm/MountainDewritoes/MountainDewritoes.java
@@ -47,7 +47,6 @@ import net.sacredlabyrinth.phaed.simpleclans.managers.ClanManager;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.GameRule;
-import org.bukkit.NamespacedKey;
 import org.bukkit.SoundCategory;
 import org.bukkit.World;
 import org.bukkit.configuration.InvalidConfigurationException;
@@ -70,7 +69,6 @@ import org.bukkit.metadata.Metadatable;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.RegisteredServiceProvider;
 import org.bukkit.plugin.java.JavaPlugin;
-import org.bukkit.persistence.PersistentDataType;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import protocolsupport.api.ProtocolSupportAPI;
@@ -135,36 +133,11 @@ public class MountainDewritoes extends JavaPlugin implements Listener
     private TipCommand tipCommand;
     private AtmosphericManager atmosphericManager;
     private CustomItemRegistry customItemRegistry;
+    private PlayerDataStore playerDataStore;
 
-    private NamespacedKey nameColorKey;
-    private NamespacedKey expLevelKey;
-
-    public ChatColor getNameColor(Player player)
+    public PlayerDataStore getPlayerDataStore()
     {
-        String color = player.getPersistentDataContainer().get(nameColorKey, PersistentDataType.STRING);
-        if (color != null)
-            return ChatColor.valueOf(color);
-
-        int colorCode = Math.abs(player.getUniqueId().hashCode());
-        String[] acceptableColors = "2,3,4,5,6,9,a,b,c,d,e".split(",");
-        colorCode = colorCode % acceptableColors.length;
-        return ChatColor.getByChar(acceptableColors[colorCode]);
-    }
-
-    public void setNameColor(Player player, ChatColor color)
-    {
-        player.getPersistentDataContainer().set(nameColorKey, PersistentDataType.STRING, color.name());
-    }
-
-    public int getStoredExpLevel(Player player)
-    {
-        Integer level = player.getPersistentDataContainer().get(expLevelKey, PersistentDataType.INTEGER);
-        return level == null ? 0 : level;
-    }
-
-    public void setStoredExpLevel(Player player, int level)
-    {
-        player.getPersistentDataContainer().set(expLevelKey, PersistentDataType.INTEGER, level);
+        return playerDataStore;
     }
 
     @Deprecated
@@ -479,8 +452,7 @@ public class MountainDewritoes extends JavaPlugin implements Listener
 
         //Classes other classes might want to use
         new NSA(this);
-        nameColorKey = new NamespacedKey(this, "name_color");
-        expLevelKey = new NamespacedKey(this, "exp_level");
+        playerDataStore = new PlayerDataStore(this);
 
         //Wow, lots-o-listeners
         PluginManager pm = getServer().getPluginManager();
